Animations fully working
Entity class has a new (and old) Direction enum, with DIRECTIONX_MOVE and DIRECTIONX_IDLE values, a _dir property and set and get methods.

Entity now as an Animator, initialized on constructor with the passed &texture.-mPlayer has a new initAnimations method that creates and addFrames to every animation it will use in the game. On update call, it chooses the correct animation according to the Event and calls update on the animator.