Added States, the base class, GameState and SettingsState
Game now creates States, and has a stackof them, to keep updating the top. GameState is started right on the Game constructor and set to the top. When pressing TAB on GameState, the SettingsState is invoked to display FPS data in the top-left screen corner.

Inputs, updates and render sequences reworkeds to call States. Entities moved to GameState.